add a class ziphandler
